feat(version): separate a new module Version
